Move regression test into v2.CreateNamespaceSuite
Replace the standalone CreateNamespaceExecSuite with:
- ValidatingInMemoryTableCatalog: an InMemoryTableCatalog subclass whose
  createNamespace validates before checking existence and raises a
  non-NamespaceAlreadyExistsException on a pre-existing namespace.
- Registration in v2.CommandSuiteBase as `validating_test_catalog`.
- One SQL-level test in v2.CreateNamespaceSuite that creates a namespace
  and re-runs CREATE NAMESPACE IF NOT EXISTS — exercises the recovery
  path end-to-end through the analyzer/planner/exec.

Generated-by: Claude Code (Opus 4.7)
